refactor: UserBalance 도메인
- enums 패키지 평면화
- 상수 표현 수정
