+++ src/main/java/kr/hhplus/be/server/domain/balance/UserBalanceService.java
@@ -1,6 +1,9 @@
 package kr.hhplus.be.server.domain.balance;
 
-import kr.hhplus.be.server.domain.balance.enums.TransactionType;
+import org.springframework.orm.ObjectOptimisticLockingFailureException;
+import org.springframework.retry.annotation.Backoff;
+import org.springframework.retry.annotation.EnableRetry;
+import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -8,7 +11,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-@Transactional
+@EnableRetry
 public class UserBalanceService {
 
     private final UserBalanceRepository userBalanceRepository;
@@ -27,12 +30,14 @@ public class UserBalanceService {
         return UserBalanceInfo.from(userBalance);
     }
 
+    @Transactional
     public UserBalanceInfo charge(UserBalanceCommand.Charge command) {
         UserBalance userBalance = userBalanceRepository.getByUserId(command.getUserId());
 
         long beforePoint = userBalance.getBalance();
         UserBalance chargedUserBalance = userBalance.charge(command.getAmount());
         long afterPoint = chargedUserBalance.getBalance();
+        userBalanceRepository.save(chargedUserBalance);
 
         UserBalanceHistory history = UserBalanceHistory.create(
                 command.getUserId(),
@@ -41,19 +46,19 @@ public class UserBalanceService {
                 beforePoint,
                 afterPoint
         );
-
         userBalanceHistoryRepository.save(history);
-        userBalanceRepository.save(userBalance);
 
-        return UserBalanceInfo.from(userBalance);
+        return UserBalanceInfo.from(chargedUserBalance);
     }
 
+    @Transactional
     public UserBalanceInfo use(UserBalanceCommand.Use command) {
         UserBalance userBalance = userBalanceRepository.getByUserId(command.getUserId());
 
         long beforePoint = userBalance.getBalance();
         UserBalance usedUserBalance = userBalance.use(command.getAmount());
         long afterPoint = usedUserBalance.getBalance();
+        userBalanceRepository.save(userBalance);
 
         UserBalanceHistory history = UserBalanceHistory.create(
                 command.getUserId(),
@@ -62,9 +67,7 @@ public class UserBalanceService {
                 beforePoint,
                 afterPoint
         );
-
         userBalanceHistoryRepository.save(history);
-        userBalanceRepository.save(userBalance);
 
         return UserBalanceInfo.from(userBalance);
     }
